+++ mockup/patterns/pickadate/pattern.js
@@ -102,7 +102,14 @@ define([
         selectYears: true,
         selectMonths: true,
         formatSubmit: 'yyyy-mm-dd',
-        format: 'yyyy-mm-dd'
+        format: 'yyyy-mm-dd',
+        clear: _t('Clear'),
+        close: _t('Close'),
+        today: _t('Today'),
+        labelMonthNext: _t('Next month'),
+        labelMonthPrev: _t('Previous month'),
+        labelMonthSelect: _t('Select a month'),
+        labelYearSelect: _t('Select a year')
       },
       time: {
       },